feat: scan finalized blocks only (#12)
To eliminate the impact caused by l2-reorg, the l2-scanner choose to
only scan the finalized blocks. Once a l2-block has been finalized, it
can be confirmed that it will not be reorg.

Previously, we configured Misc.ConfirmBlocks to eliminate L2 reorg impact, but
finalized block tags are more robust. So this PR removes
Misc.ConfirmBlocks also.